+++ sdk/LoginRadius-JavaSDK/src/main/java/com/loginradius/sdk/models/configuration/PasswordLength.java
@@ -1,0 +1,8 @@
+
+package com.loginradius.sdk.models.configuration;
+
+
+public class PasswordLength {
+
+
+}