+++ sdk/LoginRadius-JavaSDK/src/main/java/com/loginradius/sdk/models/configuration/Provider.java
@@ -1,10 +1,8 @@
 
-package com.loginradius.sdk.models.socialinterface;
-
+package com.loginradius.sdk.models.configuration;
 
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
-
 
 public class Provider {
 
@@ -15,38 +13,18 @@ public class Provider {
     @Expose
     private String endpoint;
 
-    /**
-     * 
-     * @return
-     *     The name
-     */
     public String getName() {
         return name;
     }
 
-    /**
-     * 
-     * @param name
-     *     The Name
-     */
     public void setName(String name) {
         this.name = name;
     }
 
-    /**
-     * 
-     * @return
-     *     The endpoint
-     */
     public String getEndpoint() {
         return endpoint;
     }
 
-    /**
-     * 
-     * @param endpoint
-     *     The Endpoint
-     */
     public void setEndpoint(String endpoint) {
         this.endpoint = endpoint;
     }