+++ sdk/LoginRadius-JavaSDK/src/main/java/com/loginradius/sdk/models/configuration/Question.java
@@ -1,0 +1,32 @@
+
+package com.loginradius.sdk.models.configuration;
+
+import com.google.gson.annotations.Expose;
+import com.google.gson.annotations.SerializedName;
+
+public class Question {
+
+    @SerializedName("QuestionId")
+    @Expose
+    private String questionId;
+    @SerializedName("Question")
+    @Expose
+    private String question;
+
+    public String getQuestionId() {
+        return questionId;
+    }
+
+    public void setQuestionId(String questionId) {
+        this.questionId = questionId;
+    }
+
+    public String getQuestion() {
+        return question;
+    }
+
+    public void setQuestion(String question) {
+        this.question = question;
+    }
+
+}